requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

motor==3.3.2
pymongo==4.6.0

celery==5.3.4
redis==5.0.1

openai==1.3.7
httpx==0.25.2
pillow==11.1.0
pillow-avif-plugin==1.4.6
pillow-heif==0.21.0
requests==2.31.0

python-dateutil==2.8.2

python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
jinja2==3.1.4
boto3==1.34.0
gunicorn

beautifulsoup4==4.12.3
lxml==5.1.0
tweepy==4.14.0

apscheduler==3.10.4
certifi
